# openapi.yaml
openapi: 3.0.3
info:
  title: 邻里知光 API
  version: 1.0.0
  description: |
    基于《邻里知光-技术分档-详细版》与补充 docx 整理的目标态接口文档。
    当前工作区无代码快照，本规范用于联调、评审与后续 Swagger/Knife4j/Postman 导入。
servers:
  - url: /api/v1
    description: 默认 API 前缀
security:
  - BearerAuth: []
...
paths:
...
  /trade/activities:
    get:
      tags: [Trade]
      summary: 查询交易活动列表
      operationId: listTradeActivities
      x-maturity: 已实现
      x-source: 当前代码
      security: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/TradeActivityListData'
        '400':
          $ref: '#/components/responses/ValidationError'
    post:
      tags: [Trade]
      summary: 创建交易活动
      operationId: createTradeActivity
      x-maturity: 已实现（基础版）
      x-source: 当前代码
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TradeCreateActivityRequest'
      responses:
        '200':
          description: 创建成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/TradeActivityData'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '409':
          $ref: '#/components/responses/ConflictError'
  /trade/activities/{activityId}:
    get:
      tags: [Trade]
      summary: 查询交易活动详情
      operationId: getTradeActivity
      x-maturity: 已实现
      x-source: 当前代码
      security: []
      parameters:
        - name: activityId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
        - name: quantity
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
          description: 下单数量，默认 1
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/TradeActivityData'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /trade/activities/{activityId}/orders:
    post:
      tags: [Trade]
      summary: 预扣库存并受理下单
      operationId: placeTradeOrder
      x-maturity: 已实现（基础版）
      x-source: 当前代码
      parameters:
        - name: activityId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        '200':
          description: 受理成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/TradeSubmitData'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '429':
          $ref: '#/components/responses/TooManyRequestsError'
  /trade/orders/me:
    get:
      tags: [Trade]
      summary: 查询当前用户订单列表
      operationId: listMyTradeOrders
      x-maturity: 已实现
      x-source: 当前代码
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - name: status
          in: query
          schema:
            type: string
            enum: [PENDING_PAYMENT, PAID, CLOSED]
          description: 可选订单状态筛选
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/TradeOrderPageData'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /trade/orders/{orderNo}:
    get:
      tags: [Trade]
      summary: 查询当前用户订单详情
      operationId: getMyTradeOrder
      x-maturity: 已实现
      x-source: 当前代码
      parameters:
        - name: orderNo
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/TradeOrderData'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /trade/orders/{orderNo}/status:
    get:
      tags: [Trade]
      summary: 查询当前用户订单受理状态
      operationId: getMyTradeOrderStatus
      x-maturity: 已实现
      x-source: 当前代码
      parameters:
        - name: orderNo
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/TradeOrderStatusData'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /trade/orders/{orderNo}/cancel:
    post:
      tags: [Trade]
      summary: 取消当前用户自己的未支付订单
      operationId: cancelTradeOrder
      x-maturity: 宸插疄鐜帮紙鍩虹鐗堬級
      x-source: 褰撳墠浠ｇ爜
      parameters:
        - name: orderNo
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 鍙栨秷鎴愬姛
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/TradeOrderData'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
  /trade/orders/{orderNo}/pay:
    post:
      tags: [Trade]
      summary: 模拟支付订单
      operationId: payTradeOrder
      x-maturity: 已实现（基础版）
      x-source: 当前代码
      parameters:
        - name: orderNo
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TradePayRequest'
      responses:
        '200':
          description: 支付成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/TradeOrderData'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    SizeParam:
      name: size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...